requirements.txt
pyswisseph
hijri-converter
hdate
reportlab
openpyxl
stripe
python-dotenv
fastapi
uvicorn[standard]
anthropic
httpx
